requirements.txt
discord.py==1.3.3

asyncio==3.4.3
python-dateutil==2.6.1
humanize==0.5.1
aiohttp==3.6.2